themes: light, dark
Theming is now done with CSS variables, not SCSS variables. This is done
for fun, and to get some experience with CSS vars. They don't work in
IE, but none of my visitors should be using it, and in the worst case
scenario the site just loses some colors.

This commit also adds a dark theme. The light theme is selected by
default.

Other minor changes:

* intra-page hash-links are visually prefixed with '#', using a
  pseudoelement
* improved a few colors
* fixed a few minor bugs