better external links
Markdown processing now ensures that external links are rendered with
the various "target blank" attributes and an external link icon.

The external link icon is now implemented as inline SVG, which makes it
capable of inheriting the HTML colors and therefore compatible with
themes.

Also changed "optimisation" to "optimization" in a few places. MacOS's
"American English" voice-over was unable to pronounce it correctly.